Reorganise gradle project with separate modules for core and command line.
