+++ derivation-cli/src/main/java/uk/gov/register/derivation/cli/Registrar.java
@@ -1,0 +1,12 @@
+package uk.gov.register.derivation.cli;
+
+import com.amazonaws.services.s3.AmazonS3;
+import com.amazonaws.services.s3.AmazonS3Client;
+import com.google.inject.AbstractModule;
+
+public class Registrar extends AbstractModule {
+    @Override
+    protected void configure() {
+        bind(AmazonS3.class).to(AmazonS3Client.class);
+    }
+}
