+++ derivation-cli/src/test/resources/bad-countries.rsf
