+++ derivation-cli/src/test/resources/countries.rsf
