+++ derivation-cli/src/test/resources/invalid.rsf
