+++ derivation-core/build.gradle
@@ -1,0 +1,11 @@
+apply plugin: 'java'
+
+dependencies {
+    compile 'com.fasterxml.jackson.core:jackson-databind:2.8.6'
+    compile 'com.amazonaws:aws-java-sdk-s3:1.11.77'
+    compile 'commons-codec:commons-codec:1.10'
+    testCompile 'org.skyscreamer:jsonassert:1.4.0'
+    testCompile 'junit:junit:4.12'
+    testCompile 'org.hamcrest:hamcrest-library:1.3'
+    testCompile 'org.mockito:mockito-core:1.9.5'
+}
