+++ derivation-core/src/main/java/uk/gov/register/derivation/core/SerializedRegisterParseException.java
@@ -1,0 +1,11 @@
+package uk.gov.register.derivation.core;
+
+public class SerializedRegisterParseException extends RuntimeException {
+    public SerializedRegisterParseException(String message) {
+        super(message);
+    }
+
+    public SerializedRegisterParseException(String message, Exception e){
+        super(message, e);
+    }
+}
