+++ derivation-core/src/test/resources/bad-countries.rsf
@@ -1,0 +1,4 @@
+add-item	{"citizen-names":"Czech","country":"CZ","name":"Czech Republic","official-name":"The Czech Republic","start-date":"1993-01-01"}
+add-item	{"citizen-names":"Czech","country":"CZ","name":"Czechia","official-name":"The Czech Republic","start-date":"1993-01-01"}
+append-entry	2016-04-05T13:23:05Z	sha-256:c45bd0b4785680534e07c627a5eea0d2f065f0a4184a02ba2c1e643672c3f2ed	CZ
+append-entry	2016-11-11T16:25:07Z	sha-256:c69c04fff98c59aabd739d43018e87a25fd51a00c37d100721cc68fa9003a720-I-am-wrong	CZ
