+++ derivation-core/src/test/resources/invalid.rsf
@@ -1,0 +1,2 @@
+add-item	{"citizen-names":"Vatican citizen","country":"VA","name":"Vatican City","official-name":"The Vatican City"}
+append-entry	2016-04-05T13:23:05Z	sha-256:d97d6b34bc572e334cbd7898f785b72947557d9dbea59977077f231274259f3b
