--- external-dependencies.gradle
@@ -1,12 +1,0 @@
-ext {
-    googleGuice = 'com.google.inject:guice:4.1.0'
-    jackson =  'com.fasterxml.jackson.core:jackson-databind:2.8.6'
-    commonsCodec = 'commons-codec:commons-codec:1.10'
-    awsS3 = ['com.amazonaws:aws-java-sdk-s3:1.11.77', "com.fasterxml.jackson.dataformat:jackson-dataformat-cbor:2.8.6"]
-
-
-    // test dependencies
-    junit = ['junit:junit:4.12','org.hamcrest:hamcrest-library:1.3']
-    mockito = 'org.mockito:mockito-core:1.9.5'
-    jsonAssert ='org.skyscreamer:jsonassert:1.4.0'
-}
