+++ settings.gradle
@@ -1,2 +1,4 @@
 rootProject.name = 'openregister-derivation'
+include 'derivation-core'
+include 'derivation-cli'
