--- src/main/java/uk/gov/register/derivation/PartialEntity.java
@@ -1,32 +1,0 @@
-package uk.gov.register.derivation;
-
-import java.util.LinkedList;
-import java.util.List;
-
-public class PartialEntity {
-
-    private final String key;
-
-    private final List<Entry> entries;
-
-    public PartialEntity(String key) {
-        this.key = key;
-        this.entries = new LinkedList<>();
-    }
-
-    public String getKey() {
-        return key;
-    }
-
-    public List<Entry> getEntries() {
-        return entries;
-    }
-
-    @Override
-    public String toString() {
-        return "PartialEntity{" +
-                "key='" + key + '\'' +
-                ", entries=" + entries +
-                '}';
-    }
-}
