--- src/main/java/uk/gov/register/derivation/SerializedRegisterParseException.java
@@ -1,11 +1,0 @@
-package uk.gov.register.derivation;
-
-public class SerializedRegisterParseException extends RuntimeException {
-    public SerializedRegisterParseException(String message) {
-        super(message);
-    }
-
-    public SerializedRegisterParseException(String message, Exception e){
-        super(message, e);
-    }
-}
